array: made header-only, shared<T>: exposed  to public API
